added new env variable for redis , refactored utils folder , added client for redis
