+++ .env.example
@@ -3,3 +3,4 @@ SECRET_KEY=abcdefghijklmnopqrstuvwxyz1212131313
 GMAIL_ID=
 GMAIL_PASS=
 ROOT_URL=http://localhost:8082
+REDIS_URL=
