+++ package.json
@@ -54,6 +54,7 @@
     "passport": "^0.6.0",
     "passport-google-oauth20": "^2.0.0",
     "randomstring": "^1.3.0",
+    "redis": "^4.7.0",
     "shortid": "^2.2.16",
     "socket.io": "^4.7.2"
   },
